feat(notes-app-backend): authorization implemented. only Users test is passed

+++ Belajar Back-End Pemula dengan Javascript/notes-app-backend/src/services/notes/controllers/note-controller.js
@@ -1,11 +1,12 @@
-import { InvariantError, NotFoundError } from '../../../exceptions/index.js';
+import { InvariantError, NotFoundError, AuthorizationError } from '../../../exceptions/index.js';
 import response from '../../../utils/response.js';
 import NoteRepositories from '../repositories/note-repositories.js';
 
 export const createNote = async (req, res, next) => {
   const { title = 'untitled', tags, body } = req.validated;
+  const { id: owner } = req.user;
   const note = await NoteRepositories.createNote({
-    title, body, tags
+    title, body, tags, owner
   });
 
   if (!note) {
@@ -16,13 +17,20 @@ export const createNote = async (req, res, next) => {
 };
 
 export const getNotes = async (req, res) => {
-  const notes = await NoteRepositories.getNotes();
-  return response(res, 200, 'success', notes);
+  const { id: owner } = req.user;
+  const notes = await NoteRepositories.getNotes(owner);
+  return response(res, 200, 'Catatan sukses ditampilkan', notes);
 };
 
 export const getNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
   const note = await NoteRepositories.getNoteById(id);
+
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
 
   if (!note) {
     return next(new NotFoundError('Catatan tidak ditemukan'));
@@ -35,6 +43,12 @@ export const editNoteById = async (req, res, next) => {
   const { id } = req.params;
   const { title, body, tags } = req.validated;
 
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
+
   const note = await NoteRepositories.editNoteById({ id, title, body, tags });
 
   if (!note) {
@@ -46,6 +60,11 @@ export const editNoteById = async (req, res, next) => {
 
 export const deleteNoteById = async (req, res, next) => {
   const { id } = req.params;
+  const { id: owner } = req.user;
+  const isOwner = await NoteRepositories.verifyNoteOwner(id, owner);
+  if (!isOwner) {
+    return next(new AuthorizationError('Anda tidak berhak mengakses resource ini'));
+  }
   const deletedNote = await NoteRepositories.deleteNote(id);
 
   if (!deletedNote) {
